SWC-6412 - Input component theme changes
- Move StyledFormControl to SRC
- Apply most StyledFormControl > InputBase styles to all InputBase components in the theme
- Use StyledFormControl in custom TextField component

For input fields, SageAccountWeb uses StyledFormControl with an inner MUI TextField. This is flawed because TextField actually contains its own FormControl component. The input fields look acceptable for now, but we should aim to fix this in the future.
