+++ packages/synapse-react-client/src/lib/containers/TextField.tsx
@@ -1,12 +1,11 @@
 import React, { useId, useMemo } from 'react'
 import {
   Box,
-  FormControl,
   InputBase,
+  InputLabel,
   TextFieldProps as MuiTextFieldProps,
-  Typography,
-  useTheme,
 } from '@mui/material'
+import StyledFormControl from '../components/styled/StyledFormControl'
 
 type TextFieldProps = Pick<
   MuiTextFieldProps,
@@ -25,6 +24,7 @@ type TextFieldProps = Pick<
   | 'onChange'
   | 'placeholder'
   | 'required'
+  | 'rows'
   | 'sx'
   | 'type'
   | 'value'
@@ -35,8 +35,7 @@ type TextFieldProps = Pick<
  */
 export default function TextField(props: TextFieldProps) {
   const id = useId()
-  const { palette } = useTheme()
-  const { noWrapInFormControl, ...rest } = props
+  const { noWrapInFormControl, label, ...rest } = props
   const Wrapper = useMemo(
     () =>
       noWrapInFormControl
@@ -44,19 +43,18 @@ export default function TextField(props: TextFieldProps) {
             <React.Fragment>{props.children}</React.Fragment>
           )
         : (props: React.PropsWithChildren<object>) => (
-            <FormControl fullWidth sx={{ my: 1 }}>
+            <StyledFormControl fullWidth sx={{ my: 1 }}>
               {props.children}
-            </FormControl>
+            </StyledFormControl>
           ),
     [noWrapInFormControl],
   )
   return (
     <Wrapper>
       {props.label && (
-        <Typography
+        <InputLabel
           component={'label'}
           htmlFor={props.id || id}
-          variant={'body1'}
           sx={{ fontWeight: 700, mb: '4px' }}
         >
           {props.label}
@@ -67,22 +65,9 @@ export default function TextField(props: TextFieldProps) {
           ) : (
             <></>
           )}
-        </Typography>
+        </InputLabel>
       )}
-      <InputBase
-        {...rest}
-        id={props.id || id}
-        sx={{
-          backgroundColor: 'grey.200',
-          px: 2,
-          py: 1,
-          ...props.sx,
-          '&.Mui-focused': {
-            backgroundColor: palette.background.default,
-            outline: `1px solid ${palette.primary.main}`,
-          },
-        }}
-      ></InputBase>
+      <InputBase id={id} {...rest}></InputBase>
     </Wrapper>
   )
 }
